Remove DataFusion tests, update docs
