--- ballista/tests/test_dataframe.py
@@ -1,198 +1,0 @@
-# Licensed to the Apache Software Foundation (ASF) under one
-# or more contributor license agreements.  See the NOTICE file
-# distributed with this work for additional information
-# regarding copyright ownership.  The ASF licenses this file
-# to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance
-# with the License.  You may obtain a copy of the License at
-#
-#   http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing,
-# software distributed under the License is distributed on an
-# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-# KIND, either express or implied.  See the License for the
-# specific language governing permissions and limitations
-# under the License.
-
-import pyarrow as pa
-import pytest
-
-from ballista import functions as f
-from ballista import DataFrame, SessionContext, column, literal, udf
-
-
-@pytest.fixture
-def df():
-    ctx = SessionContext()
-
-    # create a RecordBatch and a new DataFrame from it
-    batch = pa.RecordBatch.from_arrays(
-        [pa.array([1, 2, 3]), pa.array([4, 5, 6])],
-        names=["a", "b"],
-    )
-
-    return ctx.create_dataframe([[batch]])
-
-
-@pytest.fixture
-def struct_df():
-    ctx = SessionContext()
-
-    # create a RecordBatch and a new DataFrame from it
-    batch = pa.RecordBatch.from_arrays(
-        [pa.array([{"c": 1}, {"c": 2}, {"c": 3}]), pa.array([4, 5, 6])],
-        names=["a", "b"],
-    )
-
-    return ctx.create_dataframe([[batch]])
-
-
-def test_explain_string(df):
-    df = df.select(
-        column("a") + column("b"),
-        column("a") - column("b"),
-    )
-
-    # execute and collect the first (and only) batch
-    df.collect()
-
-    explain = df.explain_string()
-    assert "logical_plan" in explain
-    assert "physical_plan" in explain
-
-
-def test_select(df):
-    df = df.select(
-        column("a") + column("b"),
-        column("a") - column("b"),
-    )
-
-    # execute and collect the first (and only) batch
-    result = df.collect()[0]
-
-    assert result.column(0) == pa.array([5, 7, 9])
-    assert result.column(1) == pa.array([-3, -3, -3])
-
-
-def test_filter(df):
-    df = df.select(
-        column("a") + column("b"),
-        column("a") - column("b"),
-    ).filter(column("a") > literal(2))
-
-    # execute and collect the first (and only) batch
-    result = df.collect()[0]
-
-    assert result.column(0) == pa.array([9])
-    assert result.column(1) == pa.array([-3])
-
-
-def test_sort(df):
-    df = df.sort(column("b").sort(ascending=False))
-
-    table = pa.Table.from_batches(df.collect())
-    expected = {"a": [3, 2, 1], "b": [6, 5, 4]}
-
-    assert table.to_pydict() == expected
-
-
-def test_limit(df):
-    df = df.limit(1)
-
-    # execute and collect the first (and only) batch
-    result = df.collect()[0]
-
-    assert len(result.column(0)) == 1
-    assert len(result.column(1)) == 1
-
-
-def test_udf(df):
-    # is_null is a pa function over arrays
-    is_null = udf(
-        lambda x: x.is_null(),
-        [pa.int64()],
-        pa.bool_(),
-        volatility="immutable",
-    )
-
-    df = df.select(is_null(column("a")))
-    result = df.collect()[0].column(0)
-
-    assert result == pa.array([False, False, False])
-
-
-def test_join():
-    ctx = SessionContext()
-
-    batch = pa.RecordBatch.from_arrays(
-        [pa.array([1, 2, 3]), pa.array([4, 5, 6])],
-        names=["a", "b"],
-    )
-    df = ctx.create_dataframe([[batch]])
-
-    batch = pa.RecordBatch.from_arrays(
-        [pa.array([1, 2]), pa.array([8, 10])],
-        names=["a", "c"],
-    )
-    df1 = ctx.create_dataframe([[batch]])
-
-    df = df.join(df1, join_keys=(["a"], ["a"]), how="inner")
-    df = df.sort(column("a").sort(ascending=True))
-    table = pa.Table.from_batches(df.collect())
-
-    expected = {"a": [1, 2], "c": [8, 10], "b": [4, 5]}
-    assert table.to_pydict() == expected
-
-
-@pytest.mark.skip(
-    reason="https://github.com/apache/arrow-datafusion-python/issues/135"
-)
-def test_window_lead(df):
-    df = df.select(
-        column("a"),
-        f.alias(
-            f.window(
-                "lead", [column("b")], order_by=[f.order_by(column("b"))]
-            ),
-            "a_next",
-        ),
-    )
-
-    table = pa.Table.from_batches(df.collect())
-
-    expected = {"a": [1, 2, 3], "a_next": [5, 6, None]}
-    assert table.to_pydict() == expected
-
-
-def test_get_dataframe(tmp_path):
-    ctx = SessionContext()
-
-    path = tmp_path / "test.csv"
-    table = pa.Table.from_arrays(
-        [
-            [1, 2, 3, 4],
-            ["a", "b", "c", "d"],
-            [1.1, 2.2, 3.3, 4.4],
-        ],
-        names=["int", "str", "float"],
-    )
-    pa.csv.write_csv(table, path)
-
-    ctx.register_csv("csv", path)
-
-    df = ctx.table("csv")
-    assert isinstance(df, DataFrame)
-
-
-def test_struct_select(struct_df):
-    df = struct_df.select(
-        column("a")["c"] + column("b"),
-        column("a")["c"] - column("b"),
-    )
-
-    # execute and collect the first (and only) batch
-    result = df.collect()[0]
-
-    assert result.column(0) == pa.array([5, 7, 9])
-    assert result.column(1) == pa.array([-3, -3, -3])
